+++ src/testing/RastriginTest.kt
@@ -1,6 +1,5 @@
 package testing
 
-import model.Room
 import swarm.Individual
 import kotlin.math.PI
 import kotlin.math.cos
@@ -11,9 +10,9 @@ import kotlin.math.pow
  */
 class RastriginTest {
 
-    fun evaluateIndividual(individual: Individual) : Float {
+    fun evaluateIndividual(individual: Individual) : Double {
 
-        var sumFactor = 10f * individual.coords.size * 2
+        var sumFactor = 10.0 * individual.coords.size * 2
         individual.coords.forEach { (x, y) ->
             sumFactor += (x.pow(2) + y.pow(2) - 10 * cos(2 * PI * x ) - 10 * cos(2 * PI * y)).toFloat()
         }
